add spec and reorganize files
