+++ superset-frontend/src/components/ListView/index.ts
@@ -1,0 +1,4 @@
+export * from './ListView';
+export * from './types';
+
+export { default } from './ListView';
